Backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

